RRT star implementation
select start and end points.
the algorithm will create a graph using RRT*. and then show hte path from start to end point in red.

collisionBetweenLines takes 2 lines and returns 1 if there is collision and 0 if there isn't.

collisionWithObstacle uses collisionBetweenLines to show if there is a collision between the given obstacle and the line.
(obstacle is represented as a series of line (first point of first line is equal to last point of last line)

configInsideObstacle checks if a point is inside an obstacle